Update of various rules to support bionic

+++ vagrant/cookbooks/e2web/recipes/default.rb
@@ -9,7 +9,7 @@
 require 'base64'
 
 to_install = [
-    'apache2-mpm-prefork',
+    'apache2',
     'libapache2-mod-perl2',
     'build-essential'
 ]
@@ -18,30 +18,66 @@ to_install.each do |p|
   package p
 end
 
-bash "install Linux::Pid" do
-  cwd "/tmp"
-  user "root"
-  creates "/usr/local/lib/perl/5.14.2/auto/Linux/Pid/Pid.so"
-  code <<-EOH
-cd /tmp
-rm -rf Linux-Pid*
-wget "http://search.cpan.org/CPAN/authors/id/R/RG/RGARCIA/Linux-Pid-0.04.tar.gz" &>> /tmp/linux-pid.log;
-tar xzvf Linux-Pid-0.04.tar.gz &>> /tmp/linux-pid.log
-cd Linux-Pid-0.04
-perl Makefile.PL INSTALLDIRS=vendor &>> /tmp/linux-pid.log
-make install &>> /tmp/linux-pid.log
-rm -rf Linux-Pid*
-cd ..
-  EOH
+
+if node['platform'].eql? 'ubuntu'
+load_modules = ['rewrite','proxy','proxy_http','ssl']
+  load_modules.push('perl','mpm_prefork','socache_shmcb')
+
+  ['mpm_event.conf','mpm_event.load'].each do |mod|
+    file "/etc/apache2/mods-enabled/#{mod}" do
+      action "delete"
+      notifies :restart, "service[apache2]", :delayed
+    end
+  end
 end
 
-template '/etc/apache2/conf.d/everything' do
+load_modules.each do |apache_mod|
+  link "/etc/apache2/mods-enabled/#{apache_mod}.load" do
+    action "create"
+    to "../mods-available/#{apache_mod}.load"
+    link_type :symbolic
+    owner "root"
+    group "root"
+    notifies :restart, "service[apache2]", :delayed
+  end
+end
+
+directory "/etc/apache2/conf.d/" do
+  owner "www-data"
+  group "root"
+  mode 0755
+  action :create
+end
+
+
+unless node['platform'].eql? 'ubuntu'
+  bash "install Linux::Pid" do
+    cwd "/tmp"
+    user "root"
+    creates "/usr/local/lib/perl/5.14.2/auto/Linux/Pid/Pid.so"
+    code <<-EOH
+    cd /tmp
+    rm -rf Linux-Pid*
+    wget "http://search.cpan.org/CPAN/authors/id/R/RG/RGARCIA/Linux-Pid-0.04.tar.gz" &>> /tmp/linux-pid.log;
+    tar xzvf Linux-Pid-0.04.tar.gz &>> /tmp/linux-pid.log
+    cd Linux-Pid-0.04
+    perl Makefile.PL INSTALLDIRS=vendor &>> /tmp/linux-pid.log
+    make install &>> /tmp/linux-pid.log
+    rm -rf Linux-Pid*
+    cd ..
+    EOH
+  end
+end
+
+confdir = '/etc/apache2/conf.d'
+
+template "#{confdir}/everything" do
   owner "root"
   group "root"
   mode "0755"
   action "create"
   source 'everything.erb'
-  notifies :reload, "service[apache2]", :delayed
+  notifies :restart, "service[apache2]", :delayed
 end
 
 template '/etc/apache2/mod_rewrite.conf' do
@@ -50,7 +86,7 @@ template '/etc/apache2/mod_rewrite.conf' do
   mode "0755"
   action "create"
   source "mod_rewrite.conf.erb"
-  notifies :reload, "service[apache2]", :delayed
+  notifies :restart, "service[apache2]", :delayed
 end
 
 template '/etc/apache2/apache2.conf' do
@@ -59,31 +95,21 @@ template '/etc/apache2/apache2.conf' do
   mode "0755"
   action "create"
   source 'apache2.conf.erb'
-  notifies :reload, "service[apache2]", :delayed
+  notifies :restart, "service[apache2]", :delayed
   variables(node["e2web"])
 end
 
 
-template '/etc/apache2/conf.d/ssl.conf' do
+template "#{confdir}/ssl.conf" do
   owner "root"
   group "root"
   mode "0755"
   action "create"
   source 'ssl.conf.erb'
-  notifies :reload, "service[apache2]", :delayed
+  notifies :restart, "service[apache2]", :delayed
   variables(node["e2web"])
 end
 
-['rewrite','proxy','proxy_http','ssl'].each do |apache_mod|
-  link "/etc/apache2/mods-enabled/#{apache_mod}.load" do
-    action "create"
-    to "../mods-available/#{apache_mod}.load"
-    link_type :symbolic
-    owner "root"
-    group "root"
-    notifies :reload, "service[apache2]", :delayed
-  end
-end
 
 if node["e2web"]["make_snakeoil_tls_cert"]
   bash "Create E2 snakeoil certs" do
@@ -104,19 +130,41 @@ end
 
 file '/etc/logrotate.d/apache2' do
   action "delete"
-  notifies :reload, "service[apache2]", :delayed
+  notifies :restart, "service[apache2]", :delayed
 end
 
 # Also in e2cron, e2web
 logdir = "/var/log/everything"
 datelog = "`date +\\%Y\\%m\\%d\\%H`.log"
 
+if node['platform'].eql? 'ubuntu'
+  directory "/var/run/apache2/ssl_mutex" do
+    owner "www-data"
+    group "root"
+    mode 0755
+    action :create
+    notifies :restart, "service[apache2]", :delayed
+  end
+end
+
 directory logdir do
   owner "www-data"
   group "root"
   mode 0755
   action :create
+  notifies :restart, "service[apache2]", :delayed
 end
+
+
+if node['platform'].eql? 'ubuntu'
+  directory '/var/run/apache2/ssl_mutex' do
+    owner "www-data"
+    group "root"
+    mode 0755
+    action :create
+  end
+end
+
 
 cron 'log_deliver_to_s3.pl' do
   minute '5'
@@ -130,5 +178,5 @@ cron 'check for new TLS cert' do
 end
 
 service 'apache2' do
-  supports :status => true, :restart => true, :reload => true
+  supports :status => true, :restart => true, :reload => true, :stop => true
 end
